pyproject.toml
[tool.poetry]
name = "pythoneda-event-nix-flakes"
version = "0.0.1a1"
description = "Events for PythonEDA's nix-flakes domain"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-event/nix-flakes"
packages = [{include = "pythonedaeventnixflakes"}]

[tool.poetry.dependencies]
python = "^3.8"
pythoneda-base = "^0.0.1a15"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
